feat(product-analytics): add facade and watched-models allowance
First layer of the product_analytics isolation stack. Adds backend/facade/ (api.py delegating to the new backend/logic.py, models.py re-exporting Insight and InsightVariable under the watched-models allowance), registers the two model crossings in MODEL_CROSSINGS with the required architecture.md amendment, moves map_stale_to_latest and get_query_specific_instructions out of the DRF api modules into logic, points external callers (posthog/utils, exports) at the facade, and renames _get_insight_type public for its dashboards consumer. No behavior change.

Generated-By: PostHog Desktop
Task-Id: fe7a4c08-2d95-44d9-9740-8900b208a39f
