+++ products/product_analytics/backend/api/insight_suggestions.py
@@ -23,6 +23,7 @@ from posthog.models import Team
 
 from products.annotations.backend.api.annotation_context import build_annotations_block, resolve_query_date_range
 from products.product_analytics.backend.api.ai_billing import billable_ai_properties
+from products.product_analytics.backend.logic import get_query_specific_instructions
 
 logger = structlog.get_logger(__name__)
 
@@ -160,32 +161,6 @@ def summarize_insight_result(result: Any) -> Any:
             new_result[k] = summarize_insight_result(v)
         return new_result
     return result
-
-
-def get_query_specific_instructions(kind: str) -> str:
-    if kind == "TrendsQuery":
-        return (
-            "Focus on identifying significant changes in volume, growth trends, and seasonality. "
-            "Compare the current period to the start. Identify which breakdown segment (if any) is driving the trend."
-        )
-    elif kind == "FunnelsQuery":
-        return (
-            "Focus on conversion rates between steps. When there are three or more steps, name the step-to-step "
-            "transition with the largest loss. When there are only two steps (one transition), describe the single "
-            "drop-off directly without superlatives like 'the biggest' or 'the main bottleneck' — there is nothing "
-            "to compare it against. Compare conversion across breakdown segments if available."
-        )
-    elif kind == "RetentionQuery":
-        return (
-            "Focus on the retention curve shape. Identify when the drop-off stabilizes. "
-            "Compare retention rates between different cohorts or breakdown segments."
-        )
-    elif kind == "StickinessQuery":
-        return "Focus on how frequently users engage. Identify if there is a core group of power users."
-    elif kind == "LifecycleQuery":
-        return "Focus on the balance between new, returning, resurrecting, and dormant users. Identify which group is dominating the total count."
-
-    return "Focus on the most significant patterns and anomalies in the data."
 
 
 def get_insight_analysis(
