+++ products/product_analytics/backend/api/insight_variable.py
@@ -200,28 +200,3 @@ class InsightVariableViewSet(TeamAndOrgViewSetMixin, viewsets.ModelViewSet):
             SharingAccessTokenAuthentication | SharingPasswordProtectedAuthentication,
         ):
             raise PermissionDenied("Insight variables cannot be accessed via sharing authentication")
-
-
-def map_stale_to_latest(stale_variables: dict, latest_variables: list[InsightVariable]) -> dict:
-    # Keep the variables in an insight up to date based on variable code names that exist
-    current_variables = stale_variables
-    insight_variables = latest_variables
-    final_variables = {}
-
-    # Create a lookup for insight variables by code_name for quick access
-    insight_variables_by_code_name = {var.code_name: var for var in insight_variables}
-
-    # For each variable in current_variables, update with data from insight_variables if code_name matches
-    for _, v in current_variables.items():
-        code_name = v.get("code_name")
-        if code_name in insight_variables_by_code_name:
-            # Update the variable with corresponding data from insight_variables
-            matched_var = insight_variables_by_code_name[code_name]
-            # Add attributes from matched_var that can be serialized to JSON
-            final_variables[str(matched_var.id)] = {
-                **v,
-                "code_name": matched_var.code_name,
-                "variableId": str(matched_var.id),
-            }
-
-    return final_variables
